Protect against missing products
It could happen that, on a given input file, there are multiple events with
different products generated, especially if the file has been HLT generated,
where products are dynamically generated. A protection has been put in place so
that if any event does not have the required collection, the program will skip
it.